move primitive types to primitives
